+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 2.4.1 (2018-04-23)
+ 1. **Simpletools\Mvc\Router**
+    1. Added http methods overloader
+ 2. **Simpletools\Mvc\Controller**
+    1. Added http methods overloader
+ 2. **Simpletools\Terminal\Cli**
+    1. Added Terminal CLI
+    
 ### 2.3.23 (2018-03-23)
  1. **Simpletools\Cloud\Google\Storage\File**
     1. Save after Import always force upload to Storage